fix(tauri): catch audio playing errors gracefully
https://github.com/tauri-apps/tauri/issues/3478

As the sounds are triggered by a timer rather than a user interaction the sounds don't work at the moment.

One fix for now if its taking a while to merge would be to trigger the audio in rust. This could be achieved by passing the sound location or reading the data and sending it to rust to play.